1.06: Commented more interface functions added.
Functions that cater to the `list_node_t` is added to the `list_interface`.
Further, initial function prototypes for list iterators are also added.
The addition of these functions to the `list_implementeion` and the
implementation in `list.c` will be added later as will the implementation
in the concrete data type `array_list`.