feat: dlearn commits
Fixed uintptr_t not present in ubuntu.

Fixed temporarily the memory leak in `set_node_value`. Will
need to decide what to do with this.

Will need to decide whether the return value from getters should be
const void * and whether the return value should not be linked to the
original pointer.